requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.1
django-toolbelt==0.0.1
djangorestframework==3.4.6
foreman==0.9.7
futures==3.0.5
gunicorn==19.6.0
numpy==1.11.2
pandas==0.19.1
psycopg2==2.6.2
python-dateutil==2.6.0
pytz==2016.7
requests==2.12.1
requests-futures==0.9.7
six==1.10.0
static3==0.7.0
whitenoise==3.2.2
